Added the words NAME>COMPILE and COMPILE-NAME to the Forth wordlist.
The word NAME>COMPILE is the standard Forth-2012 Programming Tools
Extensions wordset (15.6.2.1909.10).

COMPILE-NAME is a non-standard word which is similar to COMPILE, but
operates on an nt. It is useful for handling non-default compilation
semantics.

In addition, the two introspection tools programs, wl-info.4th and
wl-collisions.4th, have been combined into wl-tools.4th.

	deleted:    forth-src/wl-collisions.4th
	deleted:    forth-src/wl-info.4th
	new file:   forth-src/wl-tools.4th
	modified:   src/ForthVM.cpp
	modified:   src/ForthWords.h
	modified:   src/fbc.h
	modified:   src/kforth.cpp
	modified:   src/vm32-common.s

+++ src/kforth.cpp
@@ -59,9 +59,9 @@ extern "C" {
 extern vector<WordList> Dictionary;
 extern char* C_ErrorMessages[];
 
-extern "C" int* JumpTable;
-extern "C" int* BottomOfStack;
-extern "C" int* BottomOfReturnStack;
+extern "C" long int* JumpTable;
+extern "C" long int* BottomOfStack;
+extern "C" long int* BottomOfReturnStack;
 extern "C" char TIB[];
 extern "C" {
     void echo_on(void);
